requirements.txt
# Flask and extensions
Flask==2.3.2
Flask-PyMongo==2.3.0
Flask-APScheduler==1.12.4

# MongoDB
pymongo==4.3.3

# HL7 parsing
hl7apy==1.3.4

# FHIR client
fhirclient==4.1.0

# Date and time handling
python-dateutil==2.8.2

# Environment variable management
python-dotenv==1.0.0

# WSGI server for production
gunicorn==20.1.0

# Testing
pytest==7.3.1
pytest-flask==1.2.0

# Code linting and formatting
flake8==6.0.0
black==23.3.0

# Security
Flask-Talisman==1.0.0

# Logging
loguru==0.7.0

# For handling CORS if needed
Flask-CORS==3.0.10

# For background tasks if needed
celery==5.3.1

# For caching if needed
Flask-Caching==2.0.2

# For handling forms if needed
Flask-WTF==1.1.1

# For API documentation if needed
flasgger==0.9.5

# For pagination if needed
Flask-Paginate==2022.1.8

# For database migrations if needed
Flask-Migrate==4.0.4

# HTTP library for API requests
requests==2.31.0

# SFTP handling
paramiko==3.2.0

# PDF processing
pdfplumber==0.9.0

# YAML parsing (for FHIR mapping file)
PyYAML==6.0

# JSON handling
jsonschema==4.17.3

# For timezone handling
pytz==2023.3

# For better datetime handling
arrow==1.2.3

# For XML parsing (if needed)
lxml==4.9.2

# For data manipulation and analysis (if needed)
pandas==2.0.2
numpy==1.24.3

# For encryption (if needed)
cryptography==41.0.1

# For async operations (if needed)
aiohttp==3.8.4

# For more efficient HTTP requests (if needed)
httpx==0.24.1

# For handling environment variables
environs==9.5.0

# For generating fake data (useful for testing)
Faker==18.10.1

pyccda==1.1.1
pyx12==2.3.5
lxml==4.9.2
